Attempted to add audio to the drip, chimes, and crickets
So all sounds except for the cricket will play in the game (for whatever
reason) but I can't find out how to trigger the audio clips from the
start of the particle animations since particle triggers are collision
based and there's nothing to collide with. I can either figure it out in
the morning or someone else could take a crack at it in the meantime.

Only drip has a prefab. The other two have not been created yet because
theyre not fully functional.

NOTE: all of this is in the "Craig-LightingTest" scene for now